Rework team damage handling
Fixes https://github.com/Unvanquished/gameplay/issues/10, at least
partially.

Note that the team damage settings still does apply only to the *other*
members of a team. Do we really want to keep that? IMO it makes no
sense, as if a lucifer cannon is less dangerous for your friends, it
makes sense if it's less dangerous for you. Idem for the barbs.

comes with the corresponding change in daemon.